+++ lib/src/controls/inputs/checkbox.dart
@@ -36,15 +36,9 @@ class Checkbox extends StatelessWidget {
           duration: style.animationDuration,
           curve: style.animationCurve,
           padding: style.padding,
-          decoration: BoxDecoration(
-            color: checked
-                ? style.checkedColor(state)
-                : style.uncheckedColor(state),
-            border: checked
-                ? style.checkedBorder(state)
-                : style.uncheckedBorder(state),
-            borderRadius: style.borderRadius,
-          ),
+          decoration: checked
+              ? style.checkedDecoration(state)
+              : style.uncheckedDecoration(state),
           child: AnimatedSwitcher(
             duration: style.animationDuration,
             switchInCurve: style.animationCurve,
@@ -63,18 +57,16 @@ class Checkbox extends StatelessWidget {
 }
 
 class CheckboxStyle {
-  final ButtonState<Color> checkedColor;
-  final ButtonState<Color> uncheckedColor;
+  final ButtonState<Decoration> checkedDecoration;
+  final ButtonState<Decoration> uncheckedDecoration;
+  final ButtonState<Decoration> thirdstateDecoration;
 
   final IconData icon;
   final ButtonState<Color> checkedIconColor;
   final ButtonState<Color> uncheckedIconColor;
+  final ButtonState<Color> thirdstateIconColor;
 
   final ButtonState<MouseCursor> cursor;
-
-  final ButtonState<Border> checkedBorder;
-  final ButtonState<Border> uncheckedBorder;
-  final BorderRadiusGeometry borderRadius;
 
   final EdgeInsetsGeometry padding;
   final EdgeInsetsGeometry margin;
@@ -82,48 +74,72 @@ class CheckboxStyle {
   final Duration animationDuration;
   final Curve animationCurve;
 
-  CheckboxStyle({
-    this.checkedColor,
-    this.uncheckedColor,
+  const CheckboxStyle({
+    this.checkedDecoration,
+    this.uncheckedDecoration,
+    this.thirdstateDecoration,
     this.cursor,
-    this.checkedBorder,
-    this.uncheckedBorder,
-    this.borderRadius,
     this.padding,
     this.margin,
     this.icon,
     this.checkedIconColor,
     this.uncheckedIconColor,
+    this.thirdstateIconColor,
     this.animationDuration,
     this.animationCurve,
   });
 
-  static CheckboxStyle defaultTheme([Brightness brightness]) {
-    Color disabledColor = Colors.grey[100].withOpacity(0.6);
+  static CheckboxStyle defaultTheme(Style style, [Brightness brightness]) {
+    final disabledColor = Colors.grey[100].withOpacity(0.6);
+    final radius = BorderRadius.circular(3);
     final def = CheckboxStyle(
       cursor: buttonCursor,
-      borderRadius: BorderRadius.circular(2),
-      checkedColor: (state) {
+      checkedDecoration: (state) {
+        Color color;
         if (state.isDisabled)
-          return disabledColor;
+          color = disabledColor;
         else if (state.isHovering || state.isPressing)
-          return Colors.blue[60];
+          color = Colors.blue[60];
         else
-          return Colors.blue[50];
+          color = Colors.blue[50];
+        return BoxDecoration(
+          borderRadius: radius,
+          color: color,
+          border: Border.all(style: BorderStyle.none),
+        );
       },
-      checkedBorder: (state) => Border.all(style: BorderStyle.none),
+      thirdstateDecoration: (state) {
+        Color color;
+        if (state.isDisabled)
+          color = disabledColor;
+        else if (state.isHovering || state.isPressing)
+          color = Colors.blue[60];
+        else
+          color = Colors.blue[50];
+        return BoxDecoration(
+          borderRadius: radius,
+          color: color,
+          border: Border.all(
+            width: 4.5,
+            color: Colors.blue,
+          ),
+        );
+      },
       padding: EdgeInsets.all(4),
       margin: EdgeInsets.all(4),
       icon: FluentIcons.checkbox_checked_24_filled,
-      uncheckedColor: (_) => Colors.transparent,
       animationDuration: Duration(milliseconds: 200),
       animationCurve: Curves.linear,
     );
     if (brightness == null || brightness == Brightness.light)
       return def.copyWith(CheckboxStyle(
-        uncheckedBorder: (state) => Border.all(
-          width: 0.6,
-          color: state.isDisabled ? disabledColor : Colors.black,
+        uncheckedDecoration: (state) => BoxDecoration(
+          border: Border.all(
+            width: 0.6,
+            color: state.isDisabled ? disabledColor : Colors.black,
+          ),
+          color: Colors.transparent,
+          borderRadius: radius,
         ),
         checkedIconColor: (_) => Colors.white,
         uncheckedIconColor: (state) {
@@ -133,9 +149,13 @@ class CheckboxStyle {
       ));
     else
       return def.copyWith(CheckboxStyle(
-        uncheckedBorder: (state) => Border.all(
-          width: 0.6,
-          color: state.isDisabled ? disabledColor : Colors.white,
+        uncheckedDecoration: (state) => BoxDecoration(
+          border: Border.all(
+            width: 0.6,
+            color: state.isDisabled ? disabledColor : Colors.white,
+          ),
+          color: Colors.transparent,
+          borderRadius: radius,
         ),
         checkedIconColor: (_) => Colors.white,
         uncheckedIconColor: (state) {
@@ -148,11 +168,6 @@ class CheckboxStyle {
   CheckboxStyle copyWith(CheckboxStyle style) {
     if (style == null) return this;
     return CheckboxStyle(
-      checkedBorder: style?.checkedBorder ?? checkedBorder,
-      uncheckedBorder: style?.uncheckedBorder ?? uncheckedBorder,
-      borderRadius: style?.borderRadius ?? borderRadius,
-      checkedColor: style?.checkedColor ?? checkedColor,
-      uncheckedColor: style?.uncheckedColor ?? uncheckedColor,
       margin: style?.margin ?? margin,
       padding: style?.padding ?? padding,
       cursor: style?.cursor ?? cursor,
@@ -161,6 +176,8 @@ class CheckboxStyle {
       uncheckedIconColor: style?.uncheckedIconColor ?? uncheckedIconColor,
       animationCurve: style?.animationCurve ?? animationCurve,
       animationDuration: style?.animationDuration ?? animationDuration,
+      checkedDecoration: style?.checkedDecoration ?? checkedDecoration,
+      uncheckedDecoration: style?.uncheckedDecoration ?? uncheckedDecoration,
     );
   }
 }