+++ lib/src/styles/theme.dart
@@ -30,13 +30,11 @@ class Style {
   final Color scaffoldBackgroundColor;
   final Color bottomNavigationBackgroundColor;
 
-  final AppBarStyle appBarStyle;
   final NavigationPanelStyle bottomNavigationStyle;
   final CardStyle cardStyle;
   final CheckboxStyle checkboxStyle;
   final DialogStyle dialogStyle;
   final DividerStyle dividerStyle;
-  final IconButtonStyle iconButtonStyle;
   final IconStyle iconStyle;
   final ListCellStyle listCellStyle;
   final PivotItemStyle pivotItemStyle;
@@ -47,22 +45,14 @@ class Style {
   final TooltipStyle tooltipStyle;
 
   final ButtonStyle buttonStyle;
-  final ButtonStyle compoundButtonStyle;
-  final ButtonStyle actionButtonStyle;
-  final ButtonStyle contextualButtonStyle;
-  final ButtonStyle primaryButtonStyle;
+  final IconButtonStyle iconButtonStyle;
 
-  Style({
+  const Style({
     this.accentColor,
     this.brightness,
     this.scaffoldBackgroundColor,
     this.bottomNavigationBackgroundColor,
-    this.appBarStyle,
     this.buttonStyle,
-    this.actionButtonStyle,
-    this.compoundButtonStyle,
-    this.contextualButtonStyle,
-    this.primaryButtonStyle,
     this.cardStyle,
     this.iconButtonStyle,
     this.checkboxStyle,
@@ -80,86 +70,38 @@ class Style {
   });
 
   Style build(BuildContext context) {
-    if (this.brightness == null || this.brightness == Brightness.light)
-      return Style(
-          accentColor: Colors.blue,
-          brightness: brightness ?? Brightness.light,
-          scaffoldBackgroundColor: scaffoldBackgroundColor ?? Colors.white,
-          bottomNavigationBackgroundColor:
-              bottomNavigationBackgroundColor ?? Colors.white,
-          appBarStyle: AppBarStyle.defaultTheme().copyWith(appBarStyle),
-          cardStyle: CardStyle.defaultTheme().copyWith(cardStyle),
-          buttonStyle: ButtonStyle.defaultTheme().copyWith(buttonStyle),
-          actionButtonStyle: ButtonStyle.defaultActionButtonTheme()
-              .copyWith(actionButtonStyle),
-          compoundButtonStyle: ButtonStyle.defaultCompoundButtonTheme()
-              .copyWith(compoundButtonStyle),
-          contextualButtonStyle:
-              ButtonStyle.defaultTheme().copyWith(contextualButtonStyle),
-          primaryButtonStyle: ButtonStyle.defaultPrimaryButtonTheme()
-              .copyWith(primaryButtonStyle),
-          iconButtonStyle:
-              IconButtonStyle.defaultTheme().copyWith(iconButtonStyle),
-          checkboxStyle: CheckboxStyle.defaultTheme().copyWith(checkboxStyle),
-          toggleSwitchStyle:
-              ToggleSwitchStyle.defaultTheme().copyWith(toggleSwitchStyle),
-          pivotItemStyle:
-              PivotItemStyle.defaultTheme().copyWith(pivotItemStyle),
-          iconStyle: IconStyle.defaultTheme().copyWith(iconStyle),
-          splitButtonStyle:
-              SplitButtonStyle.defaultTheme().copyWith(splitButtonStyle),
-          listCellStyle: ListCellStyle.defaultTheme().copyWith(listCellStyle),
-          dialogStyle: DialogStyle.defaultTheme().copyWith(dialogStyle),
-          tooltipStyle: TooltipStyle.defaultTheme().copyWith(tooltipStyle),
-          dividerStyle: DividerStyle.defaultTheme().copyWith(dividerStyle),
-          snackbarStyle: SnackbarStyle.defaultTheme().copyWith(snackbarStyle),
-          bottomNavigationStyle: NavigationPanelStyle.defaultTheme()
-              .copyWith(bottomNavigationStyle),
-          radioButtonStyle:
-              RadioButtonStyle.defaultTheme().copyWith(radioButtonStyle));
-    else
-      return Style(
-          accentColor: Colors.blue,
-          brightness: brightness ?? Brightness.dark,
-          scaffoldBackgroundColor: scaffoldBackgroundColor ?? Colors.grey[160],
-          appBarStyle:
-              AppBarStyle.defaultTheme(brightness).copyWith(appBarStyle),
-          cardStyle: CardStyle.defaultTheme(brightness).copyWith(cardStyle),
-          buttonStyle:
-              ButtonStyle.defaultTheme(brightness).copyWith(buttonStyle),
-          actionButtonStyle: ButtonStyle.defaultActionButtonTheme(brightness)
-              .copyWith(actionButtonStyle),
-          compoundButtonStyle:
-              ButtonStyle.defaultCompoundButtonTheme(brightness),
-          contextualButtonStyle: ButtonStyle.defaultTheme(brightness)
-              .copyWith(contextualButtonStyle),
-          primaryButtonStyle: ButtonStyle.defaultPrimaryButtonTheme(brightness)
-              .copyWith(primaryButtonStyle),
-          iconButtonStyle: IconButtonStyle.defaultTheme(brightness)
-              .copyWith(iconButtonStyle),
-          checkboxStyle:
-              CheckboxStyle.defaultTheme(brightness).copyWith(checkboxStyle),
-          toggleSwitchStyle: ToggleSwitchStyle.defaultTheme(brightness)
-              .copyWith(toggleSwitchStyle),
-          pivotItemStyle:
-              PivotItemStyle.defaultTheme(brightness).copyWith(pivotItemStyle),
-          iconStyle: IconStyle.defaultTheme(brightness).copyWith(iconStyle),
-          splitButtonStyle: SplitButtonStyle.defaultTheme(brightness)
-              .copyWith(splitButtonStyle),
-          listCellStyle:
-              ListCellStyle.defaultTheme(brightness).copyWith(listCellStyle),
-          dialogStyle:
-              DialogStyle.defaultTheme(brightness).copyWith(dialogStyle),
-          tooltipStyle:
-              TooltipStyle.defaultTheme(brightness).copyWith(tooltipStyle),
-          dividerStyle:
-              DividerStyle.defaultTheme(brightness).copyWith(dividerStyle),
-          snackbarStyle:
-              SnackbarStyle.defaultTheme(brightness).copyWith(snackbarStyle),
-          bottomNavigationStyle: NavigationPanelStyle.defaultTheme(brightness)
-              .copyWith(bottomNavigationStyle),
-          radioButtonStyle: RadioButtonStyle.defaultTheme(brightness)
-              .copyWith(radioButtonStyle));
+    return Style(
+      accentColor: Colors.blue,
+      brightness: brightness ?? Brightness.light,
+      scaffoldBackgroundColor: scaffoldBackgroundColor ?? Colors.grey[160],
+      cardStyle: CardStyle.defaultTheme(brightness).copyWith(cardStyle),
+      buttonStyle:
+          ButtonStyle.defaultTheme(this, brightness).copyWith(buttonStyle),
+      iconButtonStyle: IconButtonStyle.defaultTheme(this, brightness)
+          .copyWith(iconButtonStyle),
+      checkboxStyle:
+          CheckboxStyle.defaultTheme(this, brightness).copyWith(checkboxStyle),
+      toggleSwitchStyle: ToggleSwitchStyle.defaultTheme(this, brightness)
+          .copyWith(toggleSwitchStyle),
+      pivotItemStyle:
+          PivotItemStyle.defaultTheme(brightness).copyWith(pivotItemStyle),
+      iconStyle: IconStyle.defaultTheme(brightness).copyWith(iconStyle),
+      splitButtonStyle: SplitButtonStyle.defaultTheme(this, brightness)
+          .copyWith(splitButtonStyle),
+      listCellStyle:
+          ListCellStyle.defaultTheme(brightness).copyWith(listCellStyle),
+      dialogStyle: DialogStyle.defaultTheme(brightness).copyWith(dialogStyle),
+      tooltipStyle:
+          TooltipStyle.defaultTheme(brightness).copyWith(tooltipStyle),
+      dividerStyle:
+          DividerStyle.defaultTheme(brightness).copyWith(dividerStyle),
+      snackbarStyle:
+          SnackbarStyle.defaultTheme(brightness).copyWith(snackbarStyle),
+      bottomNavigationStyle: NavigationPanelStyle.defaultTheme(brightness)
+          .copyWith(bottomNavigationStyle),
+      radioButtonStyle: RadioButtonStyle.defaultTheme(this, brightness)
+          .copyWith(radioButtonStyle),
+    );
   }
 
   static Style fallback(BuildContext context, [Brightness brightness]) {